Restructure folders Touch up CI
