+++ .github/workflows/release.yaml
@@ -40,9 +40,17 @@ jobs:
         github_token: ${{ secrets.GITHUB_TOKEN }}
         goos: ${{ matrix.goos }}
         goarch: ${{ matrix.goarch }}
-        project_path: "."
+        project_path: "./cmd/server"
         binary_name: "looking-glass"
         extra_files: LICENSE example.config.yaml
+    - uses: wangyoucao577/go-release-action@v1
+      with:
+        github_token: ${{ secrets.GITHUB_TOKEN }}
+        goos: ${{ matrix.goos }}
+        goarch: ${{ matrix.goarch }}
+        project_path: "./cmd/cli"
+        binary_name: "lg-cli"
+        extra_files: LICENSE
 
   docker-release:
     name: Release Docker Images
