+++ pkg/errs/ssh.go
