+++ webui/svelte.config.js
@@ -14,8 +14,8 @@ const config = {
     // If your environment is not supported or you settled on a specific environment, switch out the adapter.
     // See https://kit.svelte.dev/docs/adapters for more information about adapters.
     adapter: adapter({
-      pages: "../dist",
-      assets: "../dist",
+      pages: "../cmd/server/dist",
+      assets: "../cmd/server/dist",
       fallback: "index.html",
       precompress: false,
       strict: true,
